# Use Auto ChannelImplementation