@Maithem 2019-08-02
This PR actually improves performance on linux (uses native epoll + edge triggered option) 

**With this PR**:

18808 [Thread-1] INFO  org.corfudb.testclient.SequencerFlood  - Throughput 50328.0 req/sec Latency: mean: 0.3175660582297931 ms  50%: 0.299 ms  95%: 0.505 ms  99%: 0.75 ms
21820 [Thread-1] INFO  org.corfudb.testclient.SequencerFlood  - Throughput 50456.0 req/sec Latency: mean: 0.3170378892893448 ms  50%: 0.299 ms  95%: 0.505 ms  99%: 0.737 ms
24830 [Thread-1] INFO  org.corfudb.testclient.SequencerFlood  - Throughput 50120.0 req/sec Latency: mean: 0.3192488701636682 ms  50%: 0.298 ms  95%: 0.498 ms  99%: 0.713 ms
27837 [Thread-1] INFO  org.corfudb.testclient.SequencerFlood  - Throughput 50558.0 req/sec Latency: mean: 0.31641839588417653 ms  50%: 0.298 ms  95%: 0.501 ms  99%: 0.745 ms


**Without**:

6795 [Thread-1] INFO  org.corfudb.testclient.SequencerFlood  - Throughput 47625.0 req/sec Latency: mean: 0.33730408856812527 ms  50%: 0.308 ms  95%: 0.528 ms  99%: 0.971 ms
9804 [Thread-1] INFO  org.corfudb.testclient.SequencerFlood  - Throughput 47066.0 req/sec Latency: mean: 0.33975616560572697 ms  50%: 0.306 ms  95%: 0.528 ms  99%: 1.034 ms
21846 [Thread-1] INFO  org.corfudb.testclient.SequencerFlood  - Throughput 47883.0 req/sec Latency: mean: 0.3345878100469863 ms  50%: 0.307 ms  95%: 0.522 ms  99%: 0.803 ms